% Generated by roxygen2 (4.0.1): do not edit by hand
\name{index_plan}
\alias{index_plan}
\title{Index plan}
\usage{
index_plan(formula, colnames)
}
\arguments{
\item{formula}{The index plan formula.}

\item{colnames}{The column names of the \code{data.frame} to
index.}
}
\value{
list with column names for data, text, prefixes and
identifiers.
}
\description{
Extract the index plan from a formula specification.
}
\keyword{internal}

