% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_plan}
\alias{search_plan}
\title{Search plan}
\usage{
search_plan(formula)
}
\arguments{
\item{formula}{The term prefixes formula specification.}
}
\value{
:TODO:DOCUMENTATION:
}
\description{
Extract the term prefixes
}
\keyword{internal}

