% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,xapian_database-method}
\alias{show,xapian_database-method}
\alias{show,xapian_database-methods}
\title{Brief summary of Xapian database}
\usage{
\S4method{show}{xapian_database}(object)
}
\arguments{
\item{object}{The \code{\linkS4class{xapian_database}} object.}
}
\value{
None (invisible 'NULL').
}
\description{
Brief summary of Xapian database
}
\keyword{methods}

