% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.r
\docType{methods}
\name{summary,xapian_database-method}
\alias{summary,xapian_database-method}
\alias{summary,xapian_database-methods}
\title{Summary of a Xapian database}
\usage{
\S4method{summary}{xapian_database}(object, ...)
}
\arguments{
\item{object}{The \code{\linkS4class{xapian_database}} object.}
}
\value{
None (invisible 'NULL').
}
\description{
Summary of a Xapian database
}
\keyword{methods}
