% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixes.r
\name{term_prefixes}
\alias{term_prefixes}
\title{Term Prefixes}
\usage{
term_prefixes()
}
\value{
Character vector with valid term prefixes
}
\description{
See \url{http://xapian.org/docs/omega/termprefixes}
}
\keyword{internal}
