% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{xapian_database-class}
\alias{xapian_database-class}
\title{S4 class to handle a xapian database}
\description{
Class \code{"xapian_database"}
}
\section{Slots}{

\describe{
  \item{path}{
    Path to a Xapian database.
  }
}
}
\keyword{classes}
\keyword{methods}

