% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xapr_search}
\alias{xapr_search}
\title{Search a Xapian database}
\usage{
xapr_search(path, terms, offset = 0, pagesize = 10)
}
\arguments{
\item{path}{A character vector specifying the path to one or more
Xapian databases.}

\item{terms}{Search terms}

\item{offset}{Starting point within result set. Default 0.}

\item{pagesize}{Number of records to retrieve. Default 10.}
}
\value{
list with search result
}
\description{
Search a Xapian database
}

