% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{xsearch}
\alias{xsearch}
\alias{xsearch,xapian_database,character-method}
\title{Search a Xapian database}
\usage{
xsearch(db, query, prefix = NULL, offset = 0, pagesize = 10,
  wildcard = FALSE, view = FALSE)

\S4method{xsearch}{xapian_database,character}(db, query, prefix = NULL,
  offset = 0, pagesize = 10, wildcard = FALSE, view = FALSE)
}
\arguments{
\item{db}{A \code{linkS4class{xapian_database}} object.}

\item{query}{A free-text query}

\item{prefix}{A formula specification with term prefixes. Default
NULL. See 'Details'.}

\item{offset}{Starting point within result set. Default 0.}

\item{pagesize}{Number of records to retrieve. Default 10.}

\item{wildcard}{Support searches using a trailing '*' wildcard,
which matches any number of trailing characters, so wildc* would
match wildcard, wildcarded, wildcards, wildcat, wildcats,
etc. Default is FALSE.}

\item{view}{Invoke a text editor on the result before
returning. This enables editing the result before it's
returned. The editor is only opened if there are any response
variables supplied in the prefix formula.}
}
\value{
\code{xapian_search} object with result
}
\description{
Search a Xapian database
}
\details{
The term prefixes are specified symbolically. The
prefixes has the form '~field:prefix'.
}
\keyword{methods}

