% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xsearch}
\alias{xsearch}
\title{Search a Xapian database}
\usage{
xsearch(query, path, prefix = NULL, offset = 0, pagesize = 10,
  wildcard = FALSE)
}
\arguments{
\item{query}{A free-text query}

\item{path}{A character vector specifying the path to one or more
Xapian databases.}

\item{prefix}{A formula specification with term prefixes. Default
NULL. See 'Details'.}

\item{offset}{Starting point within result set. Default 0.}

\item{pagesize}{Number of records to retrieve. Default 10.}

\item{wildcard}{Support searches using a trailing '*' wildcard,
which matches any number of trailing characters, so wildc* would
match wildcard, wildcarded, wildcards, wildcat, wildcats,
etc. Default is FALSE.}
}
\value{
\code{xapian_search} object with result
}
\description{
Search a Xapian database
}
\details{
The term prefixes are specified symbolically. The
prefixes has the form '~field:prefix'.
}

