% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.r
\docType{class}
\name{xapian_database-class}
\alias{xapian_database-class}
\title{S4 class to handle a xapian database}
\description{
Class \code{"xapian_database"}
}
\section{Slots}{

\describe{
  \item{path}{
    Path to a Xapian database.
  }
}
}

\keyword{classes}
\keyword{methods}
