% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xapr_index}
\alias{xapr_index}
\title{Index a Xapian database}
\usage{
xapr_index(path, filename, language = c("english", "en", "none", "danish",
  "da", "dutch", "nl", "english_lovins", "lovins", "english_porter", "porter",
  "finnish", "fi", "french", "fr", "german", "de", "german2", "hungarian", "hu",
  "italian", "it", "kraaij_pohlmann", "norwegian", "nb", "nn", "no",
  "portuguese", "pt", "romanian", "ro", "russian", "ru", "spanish", "es",
  "swedish", "sv", "turkish", "tr"))
}
\arguments{
\item{path}{A character vector specifying the path to a Xapian
databases. If there is already a database in the specified
directory, it will be opened. If there isn't an existing database
in the specified directory, Xapian will try to create a new empty
database there.}

\item{filename}{A character vector specifying the file(s) to
index.}

\item{language}{Either the English name for the language or the
two letter ISO639 code. Default is 'english'}
}
\description{
Index a Xapian database
}

