% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xapr_index}
\alias{xapr_index}
\title{Index}
\usage{
xapr_index(formula, data, path, language = c("none", "english", "en",
  "danish", "da", "dutch", "nl", "english_lovins", "lovins", "english_porter",
  "porter", "finnish", "fi", "french", "fr", "german", "de", "german2",
  "hungarian", "hu", "italian", "it", "kraaij_pohlmann", "norwegian", "nb",
  "nn", "no", "portuguese", "pt", "romanian", "ro", "russian", "ru", "spanish",
  "es", "swedish", "sv", "turkish", "tr"))
}
\arguments{
\item{formula}{A formula with a symbolic description of the index
plan for the columns in the data.frame. The details of the index
plan specification are given under ‘Details’.}

\item{path}{A character vector specifying the path to a Xapian
databases. If there is already a database in the specified
directory, it will be opened. If there isn't an existing database
in the specified directory, Xapian will try to create a new empty
database there.}

\item{language}{Either the English name for the language or the
two letter ISO639 code. Default is 'none'}
}
\description{
Index the content of a \code{data.frame} with the Xapian search
engine.
}
\details{
The index plan for ‘xindex’ are specified
symbolically. An index plan has the form ‘data ~ terms’ where
‘data’ is the blob of data returned from a request and ‘terms’ is
the basis for a search in Xapian. A first order term index the
text in the column as free text. A specification of the form
'first:second' indicates that the text in 'second' should be
indexed with prefix 'first'. The prefix is a short string at the
beginning of the term to indicate which field the term
indexes. Valid prefixes are: 'A' ,'D', 'E', 'G', 'H', 'I', 'K',
'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'X', 'Y'
and 'Z'. See \url{http://xapian.org/docs/omega/termprefixes} for a
list of conventional prefixes. The specification ‘first*second’ is
the same as 'second + first:second’. The prefix 'X' will create a
user defined prefix by appending the uppercase 'second' to
'X'. The prefix 'Q' will use data in the 'second' column as a
unique identifier for the document. NA values in columns to be
indexed are skipped.
}
\examples{
\dontrun{
library(jsonlite)

## Read data
filename <- system.file("extdata/NMSI_100.csv", package="xapr")
nmsi <- read.csv(filename, as.is = TRUE)

## Store all the fields for display purposes
nmsi$data <- sapply(seq_len(nrow(nmsi)), function(i) {
    as.character(toJSON(nmsi[i,]))
})

## Index the data to a temporary database
path <- tempfile(pattern="xapr-")
dir.create(path)
xindex(data ~ S*TITLE + X*DESCRIPTION + Q:id_NUMBER, nmsi, path)
}
}

