% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\docType{methods}
\name{xsearch}
\alias{xsearch}
\alias{xsearch,xapian_database,character-method}
\title{Search a Xapian database}
\usage{
xsearch(db, query, prefix = NULL, offset = 0, pagesize = 10,
  wildcard = FALSE, view = FALSE)

\S4method{xsearch}{xapian_database,character}(db, query, prefix = NULL,
  offset = 0, pagesize = 10, wildcard = FALSE, view = FALSE)
}
\arguments{
\item{db}{A \code{linkS4class{xapian_database}} object.}

\item{query}{A free-text query}

\item{offset}{Starting point within result set. Default 0.}

\item{pagesize}{Number of records to retrieve. Default 10.}

\item{wildcard}{Support searches using a trailing '*' wildcard,
which matches any number of trailing characters, so wildc* would
match wildcard, wildcarded, wildcards, wildcat, wildcats,
etc. Default is FALSE.}

\item{view}{Invoke a text editor on the result before
returning. This enables editing the result before it's
returned. The editor is only opened if there are any response
variables supplied in the prefix formula.}

\item{formula}{A formula with symbolic specification of the data
output and term prefixes. Default NULL. See 'Details'.}
}
\value{
\code{xapian_search} object with result
}
\description{
Search a Xapian database
}
\details{
The search plan for \code{xsearch} is specified
symbolically. A search plan has the form \code{data ~
field:prefix} where \code{data} is the blob of data returned from
a request as a \code{data.frame}. The columns of the
\code{data.frame} can be specified, e.g. 'col1 + col2 ~
field:prefix'. To search prefixed terms it possible to add a
human-readable prefix; for example 'data ~ title:S' to enable
queries such as 'title:xapr'.
}
\keyword{methods}

